Fixes alignment when tableViewCell accessories set.
Note that this just fixes the button display but not hiding, sliding or
disabling of the accessory.  In the Apple code the accessory moves back
and forth with the cell but I don’t know if that is possible.
This means that trying to touch the button displayed at the far right,
under an accessory, does not work.

There is also a problem that the topContentView is a little smaller when
there is an accessory defined.  This leaves a gap between the topContentView
and the right buttons when they are exposed. This should be easy to fix by
adjusting the topContentView width up to the cell width as it is moved left.