Fixes topContentView width when accessory is set.
Note that this just fixes the button display but not hiding, sliding or
disabling of the accessory.  In the Apple code the accessory moves
back and forth with the cell but I don’t know if that is possible.
This means that trying to touch the button displayed at the far right,
under an accessory, does not work because the accessory gets the
touch.